+++ server/src/test/passage-text-parsers/chapbook/inserts/inserts.test.ts
@@ -6,7 +6,7 @@ import { ETokenType } from "../../../../semantic-tokens";
 import { MockCallbacks, buildParsingState } from "../../../builders";
 import { ChapbookParsingState } from "../../../../passage-text-parsers/chapbook/chapbook-parser";
 import { buildInsertTokens } from "./insert-builders";
-import { Token } from "../../../../passage-text-parsers/types";
+import { Token } from "../../../../types";
 import * as uutEmbedPassage from "../../../../passage-text-parsers/chapbook/inserts/embed-passage";
 import * as uutLink from "../../../../passage-text-parsers/chapbook/inserts/link";
 import * as uutRevealLink from "../../../../passage-text-parsers/chapbook/inserts/reveal-link";